Move theme declaration
